requirements.txt
django
python-dotenv
google-genai
psycopg2-binary
pypdf2
